Sat, May 28, 2022, 2:47 AM +00:00

+++ credit/credit.c
@@ -14,7 +14,7 @@ int main(void)
     printf("%i\n", a1);
     int a3 = (floor((card_number % 1000)/ 100)) * 2;
 
-    int a5 = (floor((card_number % pow(N, 5)) / pow(N, 4))) * 2;
+    float a5 = (floor((card_number % pow(N, 5)) / pow(N, 4))) * 2;
 
     int a7 = (floor((card_number % pow(N, 7)) / pow(N, 6))) * 2;
